Use compute dataset_stats

+++ configs/pi05/pi05_paligemma_tron2_full_finetune.py
@@ -166,8 +166,7 @@ train_dataloader = dict(
                         state_dim=32,
                         state_key='proprio',
                         action_key='action',
-                        norm_type='quantile',
-                        output_dtype='float32'),
+                        norm_type='quantile'),
                     dict(type='PreparePromptWithState'),
                     dict[str, str | dict[str, str]](
                         type='ProcessPrompts',
@@ -261,8 +260,7 @@ inference = dict(
                 state_dim=32,
                 state_key='proprio',
                 action_key='action',
-                norm_type='quantile',
-                output_dtype='float32'),
+                norm_type='quantile'),
             dict(type='PreparePromptWithState'),
             dict[str, str | dict[str, str]](
                 type='ProcessPrompts',
